refact

+++ src/main/java/com/trybe/acc/java/datacenter/service/AplicacaoService.java
@@ -5,6 +5,7 @@ import java.util.List;
 import javax.persistence.EntityManager;
 import javax.persistence.Query;
 
+
 public class AplicacaoService implements ServiceInterface<Aplicacao, Long> {
 
   @Override
@@ -34,6 +35,9 @@ public class AplicacaoService implements ServiceInterface<Aplicacao, Long> {
     EntityManager em = emf.createEntityManager();
 
     Aplicacao toBeDeleted = em.find(Aplicacao.class, id);
+    if (toBeDeleted == null) {
+      return;
+    }
 
     em.getTransaction().begin();
     em.remove(toBeDeleted);
@@ -49,6 +53,27 @@ public class AplicacaoService implements ServiceInterface<Aplicacao, Long> {
     Query query = em.createQuery("from Aplicacao");
 
     return query.getResultList();
+    // EntityManager em = emf.createEntityManager();
+    //
+    // CriteriaBuilder cb = em.getCriteriaBuilder();
+    // CriteriaQuery<Aplicacao> cq = cb.createQuery(Aplicacao.class);
+    // Root<Aplicacao> rootEntry = cq.from(Aplicacao.class);
+    // CriteriaQuery<Aplicacao> all = cq.select(rootEntry);
+    //
+    // TypedQuery<Aplicacao> allQuery = em.createQuery(all);
+
+
+    // EntityManager em = emf.createEntityManager();
+    //
+    // Aplicacao entity = em.find(Aplicacao.class, 1L);
+    // List<Aplicacao> list = new ArrayList<>();
+    // list.add(entity);
+    // entity = em.find(Aplicacao.class, 2L);
+    // list.add(entity);
+    // entity = em.find(Aplicacao.class, 3L);
+    // list.add(entity);
+
+    // return list;
   }
 
   @Override
